refactor (ontologies endpoint): make specific class for ontology deletion
