+++ src/models/v2/ontologies/delete/delete-ontology.ts
@@ -1,0 +1,7 @@
+import { JsonObject } from "json2typescript";
+import { UpdateDeleteOntology } from "../update-delete-ontology";
+
+@JsonObject("DeleteOntology")
+export class DeleteOntology extends UpdateDeleteOntology {
+
+}
